Reformat all codes

+++ Decode/dadecoder.cpp
@@ -5,7 +5,6 @@
 #include <string>
 #include <fstream>
 #include <algorithm>
-#include <vector>
 #include "dadecoder.h"
 
 DaDecoder::DaDecoder() = default;
@@ -25,21 +24,21 @@ void DaDecoder::getFilename(const char encodeFilename[]) {
 int DaDecoder::loadFile() {
     std::ifstream ifs;
     ifs.open(filename, std::ios::in | std::ios::binary);
-    if(ifs.fail()) return 0;
+    if (ifs.fail()) return 0;
     loadUnitInFile(ifs);
     meaning = decode();
     ifs.close();
     return 1;
 }
 
-void DaDecoder::loadUnitInFile(std::ifstream& ifs) {
+void DaDecoder::loadUnitInFile(std::ifstream &ifs) {
     int filesize = findFilesize(ifs);
     buffer = new char[filesize];
     ifs.read(buffer, filesize);
     int positionInFile = 0;
     eraseUnits();
-    for(int ii = 0; ii < unitCount; ii++){
-        DaUnit unit(*(buffer + positionInFile), *(int*)(buffer + positionInFile + 1));
+    for (int ii = 0; ii < unitCount; ii++) {
+        DaUnit unit(*(buffer + positionInFile), *(int *) (buffer + positionInFile + 1));
         positionInFile += 5;
         units.push_back(unit);
     }
@@ -48,14 +47,14 @@ void DaDecoder::loadUnitInFile(std::ifstream& ifs) {
 // Question: erase units and receive next one
 void DaDecoder::eraseUnits() {
     std::vector<DaUnit>::iterator iUnit;
-    for(iUnit = units.begin(); iUnit != units.end();) {
+    for (iUnit = units.begin(); iUnit != units.end();) {
         iUnit = units.erase(iUnit);
     }
 }
 
-int DaDecoder::findFilesize(std::ifstream& ifs) {
+int DaDecoder::findFilesize(std::ifstream &ifs) {
     ifs.seekg(0, std::ifstream::end);
-    int filesize = ifs.tellg();
+    int filesize = static_cast<int>(ifs.tellg());
     ifs.seekg(0, std::ifstream::beg);
     unitCount = filesize / daUnitSize;
     return filesize;
@@ -71,7 +70,7 @@ std::string DaDecoder::textInUnits() {
     return meaning;
 }
 
-void DaDecoder::writeIntoFile(const std::string& outfile) const {
+void DaDecoder::writeIntoFile(const std::string &outfile) const {
     std::ofstream ofs(outfile);
 //    ofs.write(meaning.c_str(), meaning.size());
     ofs << meaning;
